refactor: UIView Extension 리팩토링
